Move build discoverers to their own directory
